buypage: rename Overview to PortfolioTab
